Update to use serde Update Level to use vec instead of array Update coordinates to use usize instead of i8

+++ src/levels.rs
@@ -1,6 +1,8 @@
 #![allow(unused)]
 
-use crate::Cell;
+/// original file for levels, currently unused and shifted to json files under ./levels
+
+use sliding_puzzle::Cell;
 
 const E: Cell = Cell::Empty;
 const W: Cell = Cell::Wall;
@@ -14,7 +16,7 @@ const LEVEL1_LAYOUT: [[Cell; 12]; 6] = [
     [E, E, E, E, E, W, E, E, E, W, E, E]
 ];
 
-pub const LEVEL1: (&[i8; 2], &[i8; 2], &[[Cell; 12]; 6], &str) = 
+pub const LEVEL1: (&[usize; 2], &[usize; 2], &[[Cell; 12]; 6], &str) = 
     (&[0, 0], &[4, 4], &LEVEL1_LAYOUT, "Level 1");
 
 const LEVEL_2_LAYOUT: [[Cell; 20]; 10] = [
@@ -30,7 +32,7 @@ const LEVEL_2_LAYOUT: [[Cell; 20]; 10] = [
     [E, E, E, E, E, E, E, E, E, E, E, E, E, E, E, E, E, E, E, E], 
 ];
 
-pub const LEVEL2: (&[i8; 2], &[i8; 2], &[[Cell; 20]; 10], &str) = 
+pub const LEVEL2: (&[usize; 2], &[usize; 2], &[[Cell; 20]; 10], &str) = 
     (&[0, 0], &[1, 1], &LEVEL_2_LAYOUT, "Level 2");
 
 const LEVEL_3_LAYOUT: [[Cell; 20]; 10] = [
@@ -46,5 +48,5 @@ const LEVEL_3_LAYOUT: [[Cell; 20]; 10] = [
     [W, E, E, E, E, E, E, E, E, E, E, E, E, W, E, E, E, E, E, W],
 ];
 
-pub const LEVEL3: (&[i8; 2], &[i8; 2], &[[Cell; 20]; 10], &str) = 
+pub const LEVEL3: (&[usize; 2], &[usize; 2], &[[Cell; 20]; 10], &str) = 
     (&[0, 0], &[1, 1], &LEVEL_3_LAYOUT, "Level 3");
